+++ Changelog.md
@@ -1,5 +1,6 @@
 --------------------------Changelog for v1.9.8 of SecurityCraft--------------------------
 
+- New: The Sonic Security System's functionality can now be inverted (blocks are accessible by default, and locked when the correct tune is played)
 - Fix: The exclamation mark in the Keycard Reader screen is dark
 
 --------------------------Changelog for v1.9.7 of SecurityCraft--------------------------